moving QR to its own include file
